+++ consensus/polybft/contractsapi/apex-artifacts-gen/main.go
@@ -83,7 +83,7 @@ package contractsapi
 		log.Fatal(err)
 	}
 
-	if err = common.SaveFileSafe(currentPath+"/../apex_sc_data.go", output, 0600); err != nil {
+	if err = common.SaveFileSafe(path.Join(currentPath, "/../apex_sc_data.go"), output, 0600); err != nil {
 		log.Fatal(err)
 	}
 }
